Add solution 1.7
I was not sure how to interpret the exercise, as if you were to take it
literally, you would end up with a nonsensical result, but that is what
is implemented in this commit.

An alternative would be to also move the first println() statement to
after the end of the for-loop. Another alternative would be to subtract
i from MAX_INDEX and add 2 within the println() statement inside the
loop to keep the ascending order of the indices despite solving the
exercise by implementing the "counting backward.".

The third alternative would be to make the indices correspond to the
numbers of the Fibonacci sequence and print that in reverse order;
though due to the recursive nature of the Fibonacci sequence that seems
to be out of the scope of this exercise.