Add solution 1.9
As there is no way to know (with an iterative approach) how many values
should be printed, without computing them first, a decision needs to be
made on whether to prioritize computation time or memory storage.

For the former one has to make an array far too large, with only the
computed values inside, which is what is done here. As for the latter,
one has to compute the amount of values first, then make an array just
large enough for them, and afterwards compute the values a second time
whilst storing them.

Here it is trivial, but I thought I'd point it out.